11-06-2021 update function done

+++ HMS.Domain/UserConfig.cs
@@ -12,6 +12,6 @@ namespace HMS.Domain
         public string Address { get; set; }
         public int City { get; set; }
         public int State { get; set; }
-        public int PinCode { get; set; }
+        public long PinCode { get; set; }
     }
 }
